requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
django-extensions==3.2.3
fonttools==4.53.1
idna==3.8
inflect==7.4.0
joblib==1.4.2
jsonpickle==3.3.0
kiwisolver==1.4.5
matplotlib==3.9.2
more-itertools==10.4.0
music21==9.1.0
numpy==2.1.0
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
verovio==4.3.1
webcolors==24.8.0
